Rely on default 5432 port for PostgreSQL
We're generally unlikely to run on anything different.

I've pulled out the exposed port from development.yml because I don't
think that was intended - we're unlikely to want to connect from outside
the compose environment.

As a result I don't think we generally want or need the custom port
functionality, and can just use the default everywhere.